Lay initial foundation and Podman containermanager implementation
This patch adds the first main component to the Pulpod project, the
containermanager interface which will be used by the project to spawn
container based plugins.

As this is the first main component added to the project some extras are
also being bundled in the simplest implementation possible, these
include:
- Basic logging using `go.uber.org/zap`
- Podman in Podman testing
- Config management using `github.com/knadh/koanf/v2`

The meat of the is patch is the containermanager interface, specifically
the PodmanManager implementation. This is a wrapper around
`github.com/containers/podman/v5/pkg/bindings` to allow us to spawn
container based plugins, Pulpod itself will be running in a container
with the host Podman socket being mounted.

Testing is implemented in a way to best mirror the real topology

Ubuntu provided VM from GitHub
------------------------------
Fedora Podman container "podman_host"
------------------------------
Fedora container "pulpod_test"

This means the code being tested runs inside pulpod_test running on
Fedora with access to a Podman socket and can spawn containers on
"pulpod_host"

JIRA: https://issues.redhat.com/browse/OSPRH-11828